Course_task/requirements.txt
requests
tqdm
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client